refactor(sdk): Split event_handler module into more files
